+++ packages/puppeteer-extra-plugin-block-resources/src/example.ts
