+++ packages/puppeteer-extra-plugin-block-resources/src/index.ts
