+++ packages/puppeteer-extra-plugin-block-resources/test/index.test.ts
@@ -2,7 +2,7 @@ const PLUGIN_NAME = 'block-resources'
 
 import test from 'ava'
 
-import Plugin from '.'
+import Plugin from '../src'
 
 test('is a function', async t => {
   t.is(typeof Plugin, 'function')
